+++ NEWS.md
@@ -1,8 +1,6 @@
 # purrr 0.2.4.9000
 
-* `modify()`, `modify_if()` and `modify_at()` now preserve the class of atomic
-  vectors instead of promoting them to lists. New S3 methods are provided for
-  character, logical, double, and integer classes (@t-kalinowski, #417).
+* `accumulate()` and `accumulate_right()` now inherit names from their first input (@AshesITR, #446)
 
 * `attr_getter()` no longer uses partial matching. For example,
   if an `x` object has a `labels` attribute but no `label` attribute,
@@ -17,6 +15,10 @@
   ignoring elements at those positions.
 
 * `map()` and `modify()` now work with calls and pairlists (#412).
+
+* `modify()`, `modify_if()` and `modify_at()` now preserve the class of atomic
+  vectors instead of promoting them to lists. New S3 methods are provided for
+  character, logical, double, and integer classes (@t-kalinowski, #417).
 
 * `list_modify()`, `update_list()` and `list_merge()` now handle duplicate
   duplicate argument names correctly (#441, @mgirlich).