+++ NEWS.md
@@ -23,8 +23,8 @@
 * `list_modify()`, `update_list()` and `list_merge()` now handle duplicate
   duplicate argument names correctly (#441, @mgirlich).
   
-* Support for raw vectors with added functions `map_raw`, `imap_raw`, 
-  `flatten_raw`, `invoke_map_raw`, `map2_raw`, `pmap_raw`. 
+* `map_raw`, `imap_raw`, `flatten_raw`, `invoke_map_raw`, `map2_raw` and `pmap_raw`
+  added to support raw vectors. (#455, @romainfrancois)
 
 # purrr 0.2.4