+++ NEWS.md
@@ -22,9 +22,11 @@
 
 * `list_modify()`, `update_list()` and `list_merge()` now handle duplicate
   duplicate argument names correctly (#441, @mgirlich).
-  
+
 * `map_raw`, `imap_raw`, `flatten_raw`, `invoke_map_raw`, `map2_raw` and `pmap_raw`
   added to support raw vectors. (#455, @romainfrancois)
+
+* `pwalk`, and `walk2` conserve expressions using `quote()` or `rlang::expr()` (@keithgw, #428)
 
 # purrr 0.2.4